requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.1
pypdf>=3.15.1
python-dotenv>=1.0.0
pydantic>=2.4.2
click>=8.1.7
tqdm>=4.66.1
fpdf>=1.7.2
ctransformers>=0.2.27
huggingface-hub>=0.19.4
sentence-transformers>=2.2.2
flask>=2.0.0
werkzeug>=2.0.0
flask-bootstrap>=3.3.7
